requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
paho-mqtt==1.6.1
pycryptodome==3.14.1
requests==2.27.1
tuya-connector-python==0.1.2
urllib3==1.26.9
websocket-client==1.3.2
